ci: conditional builds
